Separate AST dump tests into dedicated test file
- Create new PmdAstDumpTest.java with all 8 AST dump tests
- Remove AST dump tests from PmdWrapperTest.java to minimize changes to existing files
- All tests passing (8 AST dump tests + existing tests)
